Rename `program` to `command`
