+++ include/pisa/command/invert.hpp
