+++ src/invert.cpp
@@ -1,4 +1,3 @@
 #include "invert.hpp"
-#include "program/invert.hpp"
 
 PISA_MAIN(Invert, invert::invert_forward_index)
